requirements.txt
openai
aiogram
python-dotenv
pandas==2.2.3
gspread>=6.0.0
google-auth>=2.0.0
